long double phase shift

+++ ptimeDLib.h
@@ -10,8 +10,8 @@
 //#include "tempo2pred_int.h"
 #include <gsl/gsl_multimin.h>
 
-int deDM (int nphase, int npol, double *in, double phaseShift, double *out);
+int deDM (int nphase, int npol, double *in, long double phaseShift, double *out);
 int write_prof (subintegration *sub, pheader *header, double *profile);
-double phaseShiftDM (subintegration *sub, pheader *header, T2Predictor pred);
+long double phaseShiftDM (subintegration *sub, pheader *header, T2Predictor pred);
 int modify_freq (subintegration *sub, pheader *header);
 int createNewfile (char *input, char *output, char *ext);
