+++ ptimeTLib.h
@@ -80,7 +80,7 @@ int remove_baseline (double *in, int index, double frac_off, int n, double *out)
 int pre_diff (double *s, int nphase, int index, double frac_off, double *s_out);
 int InitialGuess (double *s, double *p, int nphase, int nchn, int *chn);
 int preA7_QUV (double *p, int nphase, double *real_p, double *ima_p);
-int rotate (int N, double *real_p, double *real_p_rotate, double *ima_p, double *ima_p_rotate, double rot);
+int rotate (int N, double *real_p, double *real_p_rotate, double *ima_p, double *ima_p_rotate, long double rot);
 int align (int N, double phase, double b, double a, double *real_p, double *real_p_align, double *ima_p, double *ima_p_align, double rotate);
 int inverse_dft (double *real_p, double *ima_p, int ncount, double *p_new);
 int allocateMemory (params *param, int nchn, int nphase);
